replaced ~ with ENV[HOME] && improved docs

+++ tasks/linkNodeModules/index.js
@@ -1,6 +1,7 @@
 // =========== [ REQUIRE ] ===========
 var co = require("co");
-var Prompt = require("dm-prompt");
+var colors = require("colors");
+var dmPrompt = require("dm-prompt");
 var _ = require("underscore");
 require("shelljs/global");
 
@@ -11,43 +12,50 @@ var job = {};
 job.start = co.wrap(function*() {
     try {
         var directoryAnswer =
-            yield Prompt({
+            yield dmPrompt({
                 type: "input",
                 name: "directory",
                 message: "Please state the Directory to where node_modules should link to:"
             });
         var directory = directoryAnswer.directory;
-        if (directory[directory.length - 1] !== "/") {
-            directory += "/";
+        if (directory[0] === "~") {
+            var home = env['HOME'];
+            directory = directory.replace("~", home);
         }
-        var commandOutput = exec("cd ~/code/dm && ls -d */", {
-            silent: true
-        }).output;
-        var directories = commandOutput.split(/\r?\n/);
-        directories.pop();
-        directories = _.filter(directories, function(directory) {
-            if (directory[0] === "_") {
-                return false;
-            } else {
-                return true;
+        if (test("-d", directory)) {
+            if (directory[directory.length - 1] !== "/") {
+                directory += "/";
             }
-        });
-        for (var i = 0, l = directories.length; i < l; i++) {
-            var v = directories[i];
-            if (v[v.length - 1] === "/") {
-                v = v.substring(0, v.length - 1);
-            }
-            var linkPath = directory + v;
+            var commandOutput = exec("cd ~/code/dm && ls -d */", {
+                silent: true
+            }).output;
+            var directories = commandOutput.split(/\r?\n/);
+            directories.pop();
+            directories = _.filter(directories, function(directory) {
+                if (directory[0] === "_") {
+                    return false;
+                } else {
+                    return true;
+                }
+            });
+            for (var i = 0, l = directories.length; i < l; i++) {
+                var v = directories[i];
+                if (v[v.length - 1] === "/") {
+                    v = v.substring(0, v.length - 1);
+                }
+                var linkPath = directory + v;
 
-            var linkCommand = "ln -s " + linkPath + " node_modules/" + v;
-            exec('rm node_modules/' + v , {
-              silent: false
-            });
-            exec(linkCommand, {
-              silent: false
-            });
+                var linkCommand = "ln -s " + linkPath + " node_modules/" + v;
+                exec('rm node_modules/' + v, {
+                    silent: true
+                });
+                exec(linkCommand, {
+                    silent: true
+                });
+            }
+        } else {
+            console.log("The path you gave me, is not existent!".red);
         }
-
     } catch (e) {
         console.log("Filename: ", __filename, "\n", e.stack);
     }
